API: only real embed URLs, reject same-host as detail (TOPHDSTREAMS/ropezara). Plugin: save/fetch fixes, timeout 60s, names from matchTitle, error wrappers.
Made-with: Cursor

+++ wp-football-live/wp-content/themes/football-live/category-news.php
@@ -2,7 +2,7 @@
 if (!defined('ABSPATH')) { exit; }
 get_header();
 $hero_bg = 'https://images.unsplash.com/photo-1579952363873-27f3bade9f55?w=1920&q=80';
-$paged = max(1, get_query_var('paged'));
+$paged = max(1, (int) get_query_var('paged'));
 ?>
 <section class="hero">
   <div class="hero-bg" style="background-image:url('<?php echo esc_url($hero_bg); ?>')"></div>
